soutenance blanche

+++ backend/controllers/AdminControllers.js
@@ -1,7 +1,7 @@
 const db = require('../middleware/connect');
 const mysql = require("mysql");
 const User = require("../models/users");
-
+const jwt = require('jsonwebtoken')
 
 exports.findAllUsers = (req, res) => {
     let sql = 'SELECT * FROM userGroupamania ';
@@ -10,23 +10,27 @@ exports.findAllUsers = (req, res) => {
        
         if (err) {
           if (err.kind === "not_found") {
-            res.status(405).send({
-              message: "Utilisateurs non trouvé"
-            });
+            res.status(405).send({ message: "Utilisateurs non trouvé" });
           } else {
-            res.status(500).send({
-              message: "Error retrieving  ",
-            });
+            res.status(500).send({ message: "Error retrieving  ",});
           }
         } else {  
-          console.log(data)
-          res.status(200).json({
-            AllUsers : data
-           })  
+          res.status(200).json({ AllUsers : data })  
         }
       });
     };
-    exports.deleteUser = (req, res) => { 
+
+    exports.deleteUser = (req, res, next) => { 
+        const token = req.headers.authorization.split(' ')[1];
+        const decodedToken = jwt.verify(token, 'Code_SECRET_Pour_Token_Randomisé');
+        const userId = decodedToken.userId;
+        const isAdmin = decodedToken.isAdmin;
+
+       
+       if (isAdmin === 0) {
+         res.status(401).json({ message: "Vous n'êtes pas Administrateur, vous ne pouvez donc pas supprimer un utilisateur"})
+       }else {
+       
         let sql =  'DELETE FROM userGroupamania WHERE id= ?';
         db.query(sql, [req.query.id] , function (err, data, fields){
           if (err) {
@@ -44,6 +48,7 @@ exports.findAllUsers = (req, res) => {
               message:  "l'utilisateur a bien était supprimé !" 
             });
           }
-        });    
-      
-        }
+        }) 
+      }
+    
+  }
